rearranged augmentation
